CDMS-260: Base64json the cookie_policy cookie (#84)
The current implementation is to stringify the JSON object which contains the analytics config value for the tracking preference cookie.
This is valid, but causes the CDP Defra ID stub to throw an exception because it does not know how to parse the cookie.

This uses Hapi's base64json encoding for the cookie instead as a workaround for the issue, which stringifies and then base64 encodes the object.