CDMS-260: Hide the cookie banner on error pages, delete cookie_policy if invalid (#87)
* CDMS-260: Hide the cookie banner on error pages, delete cookie_policy if invalid

We don't seem to have access to the state on error pages, so we can't determine if a user has accepted cookies or not.  This change makes it so that the cookie banner is never shown on error pages.

The cookie_policy cookie has also been configured to delete itself if it is invalid.  An error page is still shown briefly but if the user reloads it shows the correct response.

* CDMS-260: Update test to handle potential for multiple cookies