CDMS-882: Iterate over decisions to populate decisions column
Currently the UI iterates over the documents in a commodity to generate the info for the decisions column.
This is problematic in the case where a CHED is required by CDS to be created because there are no documents to iterate over.

This changes the logic to use the decisions array instead for the decisions column, picking up the documents where necessary.