CDMS-957: Show the quantity if one of the doc codes is C640 (#147)
* CDMS-957: Show the quantity if one of the doc codes is C640

CHEDA/C640 decisions should show the quantity, not the weight.
This updates the logic to check if one of the provided documents is a C640 and then shows the quantity.

In the case of a netMass or supplementaryUnits not being provided, for whatever reason, it falls back to the other value if available.

* CDMS-957: Update naming for readability

---------

Co-authored-by: Matthew O'Donoghue <mattyod@gmail.com>
Co-authored-by: Mohammed Shahin Uddin <mo@thetestroom.com>